Support non-recursive import trees
gomu: Deploy local changes before incrementing version from v0.5.1
